A quick-and-dirty workaround to the escaping bug.
We simply salt the hashes we use to escape the special
characters, so that the values in the escape table will be unknown
to attackers.

It may be better to use proper escaping rather than this simple
substitution (with no way to escape pre-"escaped" strings), but
this fits the bill and plugs the hole for now.